Update README.md
Changed the build instructions for Fedora 25, since F22 is no longer supported, qmake is version specific, and python-serial package was changed to pyserial.  There is still an issue with flite https://github.com/ArduPilot/apm_planner/issues/883 which I will look at later as well as try to set up a clean system to rebuild and retest.  These are just the obvious errors.